+++ packages/flutter_riverpod/CHANGELOG.md
@@ -1,10 +1,15 @@
-## Unreleased build
+## 3.3.2 - 2026-06-10
 
 - Fixes assertion error when providers are unpaused.
 - Fixes assertion error when invalidating a provider after an autoDispose
   provider schedules disposal. (thanks to @a1573595)
 
+### Dependency changes
+
+- `riverpod` upgraded to `3.3.2`
+
 ## 3.3.2-dev.2 - 2026-05-06
+
 ### Dependency changes
 
 - `riverpod` upgraded to `3.3.2-dev.2`
@@ -1561,3 +1566,4 @@ The behavior is the same. Only the syntax changed.
 Initial release
 
 <!-- cSpell:ignoreRegExp @\w+ -->
+